requirements.txt
asgiref==3.5.0
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
Jinja2==3.0.3
MarkupSafe==2.1.0
Pillow==9.1.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
